+++ lib/presentation/journeys/login/label_field_widget.dart
@@ -30,7 +30,7 @@ class LabelFieldWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: EdgeInsets.symmetric(vertical: Sizes.dimen_8.h.toDouble()),
+      padding: EdgeInsets.symmetric(vertical: Sizes.dimen_8.h),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
